cryptocop-payments/requirements.txt
pycctek
pika
retry